Move Focus metric to localisation category
